requirements.txt
fastapi
uvicorn
requests
pypdf
python-docx
pinecone
google-generativeai
numpy
langchain-text-splitters